Move most of the AngledWatercraftFrameBlock#use logic into WoodenAngledWatercraftFrameBlock#use as it makes more sense for it to be there
